requirements.txt
streamlit
ultralytics
opencv-python-headless
pillow
gtts
numpy
